api1/requirements.txt
Flask
flask-cors
tensorflow
numpy
Pillow
requests
gradio_client